Rawee

+++ ViewModels/PdVM.cs
@@ -6,9 +6,9 @@
         public string ProductName { get; set; } = null!;
         public string? CategoryName { get; set; }
         public string? BrandName { get; set; }
-        public decimal? ProductPrice { get; set; }
+        public double? ProductPrice { get; set; }
         public string? Detail { get; set; }
         public double? ProductCost { get; set; }
-        public int? ProductStock { get; set; }
+        public double? ProductStock { get; set; }
     }
 }
